requirement.txt
asgiref==3.7.2
attrs==23.1.0
build==0.10.0
CacheControl==0.12.14
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cleo==2.0.1
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.4.1
cryptography==41.0.2
defusedxml==0.7.1
distlib==0.3.7
Django==4.1.7
django-filter==21.1
django-phone-field==1.8.1
django-phonenumber-field==6.0.0
django-phonenumbers==1.0.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dulwich==0.21.5
envparse==0.2.0
filelock==3.12.2
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
itypes==1.2.0
jaraco.classes==3.3.0
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
keyring==23.13.1
lockfile==0.12.2
MarkupSafe==2.1.3
more-itertools==10.0.0
msgpack==1.0.5
oauthlib==3.2.2
packaging==23.1
pexpect==4.8.0
phonenumbers==8.13.17
Pillow==9.0.0
pkginfo==1.9.6
platformdirs==3.9.1
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
psycopg2==2.9.3
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.8.0
pyproject_hooks==1.0.0
python3-openid==3.2.0
pytz==2023.3
rapidfuzz==2.15.1
referencing==0.30.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rpds-py==0.9.2
shellingham==1.5.0.post1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
trove-classifiers==2023.7.6
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.24.2
webencodings==0.5.1
xattr==0.10.1
zipp==3.16.2
